requirements 2.txt
python-decouple
django-markdownify